+++ src/main/java/com/github/segmentio/models/Context.java
@@ -1,5 +1,7 @@
 package com.github.segmentio.models;
 
+import com.github.segmentio.Analytics;
+
 
 public class Context extends Props {
 
@@ -7,6 +9,7 @@ public class Context extends Props {
 
 	private static final String IP_KEY = "ip";
 	private static final String LIBRARY_KEY = "library";
+	private static final String LIBRARY_VERSION_KEY = "libraryVersion";
 	private static final String PROVIDERS_KEY = "providers";
 	
 	public Context() {
@@ -21,6 +24,7 @@ public class Context extends Props {
 	
 	private void addLibraryContext() {
 		this.put(LIBRARY_KEY, "analytics-java");
+		this.put(LIBRARY_VERSION_KEY, Analytics.VERSION);
 	}
 	
 	public Context setIp(String ip) {