+++ src/main/java/com/github/segmentio/models/Props.java
@@ -1,8 +1,12 @@
 package com.github.segmentio.models;
 
+import java.math.BigDecimal;
+import java.util.Collection;
 import java.util.Date;
 import java.util.HashMap;
+import java.util.Map;
 
+import org.apache.commons.lang.ClassUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -51,14 +55,30 @@ public class Props extends HashMap<String, Object> {
 		return this;
 	}
 	
+	private boolean isPrimitive (Object value) {
+		
+		boolean primitive = false;
+		
+		if (value != null) {
+			Class<?> clazz = value.getClass();
+			// http://stackoverflow.com/questions/709961/determining-if-an-object-is-of-primitive-type
+			primitive = clazz.isPrimitive() || ClassUtils.wrapperToPrimitive(clazz) != null;
+		}
+		
+		return primitive;
+	}
+	
 	public boolean allowed(Object value) {
 		
-		if (value instanceof String ||
-			value instanceof Boolean || 
-			value instanceof Integer || 
-			value instanceof Double || 
+		if (isPrimitive(value) || 
+			value instanceof String ||
 			value instanceof Date || 
-			value instanceof Props) {
+			value instanceof Props || 
+			value instanceof BigDecimal ||
+			// http://stackoverflow.com/questions/2651632/how-to-check-if-an-object-is-a-collection-type-in-java
+			value instanceof Collection || 
+			value instanceof Map ||
+			value instanceof Object[]) {
 			return true;
 		} else {
 			return false;