+++ src/test/java/com/github/segmentio/ClientTest.java
@@ -17,14 +17,12 @@ import com.github.segmentio.models.Identify;
 import com.github.segmentio.models.Track;
 
 public class ClientTest {
-
-
 	
-	private static Client client;
+	private static AnalyticsClient client;
 	
 	@BeforeClass
 	public static void setup() {
-		client = new Client("testsecret");
+		client = new AnalyticsClient("testsecret");
 	}
 
 	@Test